+++ src/RazorLight/TemplatePage.cs
@@ -1,5 +1,5 @@
-﻿using Microsoft.AspNetCore.Html;
-using RazorLight.Internal;
+﻿using RazorLight.Internal;
+using RazorLight.Text;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -33,7 +33,7 @@ namespace RazorLight
 		/// In a Razor layout page, renders the portion of a content page that is not within a named section.
 		/// </summary>
 		/// <returns>The HTML content to render.</returns>
-		protected virtual IHtmlContent RenderBody()
+		protected virtual ITemplateContent RenderBody()
 		{
 			if (BodyContent == null)
 			{
@@ -99,11 +99,11 @@ namespace RazorLight
 		/// In layout pages, renders the content of the section named <paramref name="name"/>.
 		/// </summary>
 		/// <param name="name">The name of the section to render.</param>
-		/// <returns>An empty <see cref="IHtmlContent"/>.</returns>
+		/// <returns>An empty content token.</returns>
 		/// <remarks>The method writes to the <see cref="TemplatePageBase.Output"/> and the value returned is a token
 		/// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
 		/// value does not represent the rendered content.</remarks>
-		public HtmlString RenderSection(string name)
+		public TemplateContent RenderSection(string name)
 		{
 			if (name == null)
 			{
@@ -118,11 +118,11 @@ namespace RazorLight
 		/// </summary>
 		/// <param name="name">The section to render.</param>
 		/// <param name="required">Indicates if this section must be rendered.</param>
-		/// <returns>An empty <see cref="IHtmlContent"/>.</returns>
+		/// <returns>An empty content token.</returns>
 		/// <remarks>The method writes to the <see cref="TemplatePageBase.Output"/> and the value returned is a token
 		/// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
 		/// value does not represent the rendered content.</remarks>
-		public HtmlString? RenderSection(string name, bool required)
+		public TemplateContent? RenderSection(string name, bool required)
 		{
 			if (name == null)
 			{
@@ -140,12 +140,12 @@ namespace RazorLight
 		/// </summary>
 		/// <param name="name">The section to render.</param>
 		/// <returns>
-		/// A <see cref="Task{HtmlString}"/> that on completion returns an empty <see cref="IHtmlContent"/>.
+		/// A task that on completion returns an empty content token.
 		/// </returns>
 		/// <remarks>The method writes to the <see cref="TemplatePageBase.Output"/> and the value returned is a token
 		/// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
 		/// value does not represent the rendered content.</remarks>
-		public async Task<HtmlString> RenderSectionAsync(string name)
+		public async Task<TemplateContent> RenderSectionAsync(string name)
 		{
 			if (name == null)
 			{
@@ -162,14 +162,14 @@ namespace RazorLight
 		/// <param name="required">Indicates the <paramref name="name"/> section must be registered
 		/// (using <c>@section</c>) in the page.</param>
 		/// <returns>
-		/// A <see cref="Task{HtmlString}"/> that on completion returns an empty <see cref="IHtmlContent"/>.
+		/// A task that on completion returns an empty content token.
 		/// </returns>
 		/// <remarks>The method writes to the <see cref="TemplatePageBase.Output"/> and the value returned is a token
 		/// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
 		/// value does not represent the rendered content.</remarks>
 		/// <exception cref="InvalidOperationException">if <paramref name="required"/> is <c>true</c> and the section
 		/// was not registered using the <c>@section</c> in the Razor page.</exception>
-		public Task<HtmlString?> RenderSectionAsync(string name, bool required)
+		public Task<TemplateContent?> RenderSectionAsync(string name, bool required)
 		{
 			if (name == null)
 			{
@@ -180,7 +180,7 @@ namespace RazorLight
 			return RenderSectionAsyncCore(name, required);
 		}
 
-		private async Task<HtmlString?> RenderSectionAsyncCore(string sectionName, bool required)
+		private async Task<TemplateContent?> RenderSectionAsyncCore(string sectionName, bool required)
 		{
 			if (_renderedSections.Contains(sectionName))
 			{
@@ -196,7 +196,7 @@ namespace RazorLight
 
 				// Return a token value that allows the Write call that wraps the RenderSection \ RenderSectionAsync
 				// to succeed.
-				return HtmlString.Empty;
+				return TemplateContent.Empty;
 			}
 			else if (required)
 			{